+++ python/pytauri/src/pytauri/webview.py
@@ -1,5 +1,5 @@
 """[tauri::webview](https://docs.rs/tauri/latest/tauri/webview/index.html)"""
 
-from pytauri.ffi.webview import Webview, WebviewWindow
+from pytauri.ffi.webview import Color, Cookie, SameSite, Webview, WebviewWindow
 
-__all__ = ["Webview", "WebviewWindow"]
+__all__ = ["Color", "Cookie", "SameSite", "Webview", "WebviewWindow"]